feat: nvim-tree
